requirements.txt
praw>=7.7
tweepy>=4.14
discord.py>=2.3
requests>=2.31
python-dotenv>=1.0
